src/requirements.txt
python-magic==0.4.27
libmagic==1.0
langchain==0.0.334
gradio==4.2.0
gradio_client==0.7.0
tiktoken==0.5.1
faiss-cpu==1.7.4
openai==1.2.3
unstructured==0.10.28
python-dotenv==1.0.0